app/requirements.txt
boto3==1.19.4
botocore==1.22.4
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.0
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
